feat: separate DeletePictureModal from ClubPictureItem
